Modified ply_data to actually compile
